remove length

+++ app/models.py
@@ -115,7 +115,7 @@ class Product(db.Model):
     product_title = db.Column(db.String(50), nullable=False)
     product_description = db.Column(db.String(120), nullable=False)
     # removed decimal place to test wtforms field
-    product_price = db.Column(db.Integer(12), nullable=False)
+    product_price = db.Column(db.Integer, nullable=False)
     category_category_id = db.Column(
         db.Integer, db.ForeignKey('category.category_id'))
     location_location_id = db.Column(
